tp7/tp-dvc-mlflow/pyproject.toml
[tool.poetry]
name = "tp-dvc-mlflow"
version = "0.1.0"
description = "DVC MLflow Adult Income Pipeline"
authors = ["Your Name <email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0"
scikit-learn = "^1.3"
mlflow = "^2.8"
joblib = "^1.3"
pyyaml = "^6.0"
numpy = "^1.24"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
